+++ src/main/java/io/powerrangers/backend/dao/TaskRepository.java
@@ -1,4 +1,4 @@
-package io.powerrangers.backend.repository;
+package io.powerrangers.backend.dao;
 
 import io.powerrangers.backend.entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
