--- src/main/java/io/powerrangers/backend/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package io.powerrangers.backend.repository;
-
-import io.powerrangers.backend.entity.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
