+++ src/main/java/io/powerrangers/backend/service/TaskService.java
@@ -3,8 +3,7 @@ package io.powerrangers.backend.service;
 import io.powerrangers.backend.dto.*;
 import io.powerrangers.backend.entity.Task;
 import io.powerrangers.backend.entity.User;
-import io.powerrangers.backend.repository.TaskRepository;
-import io.powerrangers.backend.repository.UserRepository;
+import io.powerrangers.backend.dao.TaskRepository;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
